+++ index.js
@@ -15,7 +15,8 @@
 const forwarded = require('@fastify/forwarded')
 const ipaddr = require('ipaddr.js')
 const { isIP } = require('net')
-
+const { isIPv4MappedIPv6Address } = require('./lib/is-ipv4-mapped-ipv6-address')
+const { prefixLengthFromSubnetMask } = require('./lib/prefix-length-from-subnet-mask')
 /**
  * Variables.
  * @private
@@ -333,115 +334,6 @@ function trustSingle (subnet) {
   }
 }
 
-function octetToZeros (ip, start, end) {
-  return (
-    (ip[start] === '0' && 9) ||
-    ((end - start) === 3 && (
-      (ip[start] === '1' && (
-        (ip[start + 1] === '2' && ip[start + 2] === '8' && 8) ||
-        (ip[start + 1] === '9' && ip[start + 2] === '2' && 7)
-      )) ||
-      (ip[start] === '2' && (
-        (ip[start + 1] === '2' && ip[start + 2] === '4' && 6) ||
-        (ip[start + 1] === '4' && (
-          (ip[start + 2] === '0' && 5) ||
-          (ip[start + 2] === '8' && 4)
-        )) ||
-        (ip[start + 1] === '5' && (
-          (ip[start + 2] === '2' && 3) ||
-          (ip[start + 2] === '4' && 2) ||
-          (ip[start + 2] === '5' && 1)
-        ))
-      ))
-    ))
-  ) || 0
-}
-
-/**
- * Parse netmask string into CIDR range.
- *
- * @param {String} netmask
- * @private
- */
-function prefixLengthFromSubnetMask (netmask) {
-  let stop = false
-  let cidr = 0
-
-  let end = netmask.length
-  let start = netmask.lastIndexOf('.', end - 1)
-  let zeros = octetToZeros(netmask, start + 1, end)
-  if (zeros === 0) return null
-  zeros !== 9 && (stop = true)
-  cidr += zeros
-
-  end = start
-  start = netmask.lastIndexOf('.', end - 1)
-  zeros = octetToZeros(netmask, start + 1, end)
-  if (zeros !== 1 && (stop === true || zeros === 0)) return null
-  stop === false && zeros !== 9 && (stop = true)
-  cidr += zeros
-
-  end = start
-  start = netmask.lastIndexOf('.', end - 1)
-  zeros = octetToZeros(netmask, start + 1, end)
-  if (zeros !== 1 && (stop === true || zeros === 0)) return null
-  stop === false && zeros !== 9 && (stop = true)
-  cidr += zeros
-
-  end = start
-  zeros = octetToZeros(netmask, 0, end)
-  if (zeros !== 1 && (stop === true || zeros === 0)) return null
-  stop === false && zeros !== 9 && (stop = true)
-  cidr += zeros
-
-  return 36 - cidr
-}
-
-function isIPv4MappedIPv6Address (addr) {
-  if (
-    addr[0] === ':' &&
-    addr[1] === ':'
-  ) {
-    if (
-      addr[2] === 'f' &&
-      addr[3] === 'f' &&
-      addr[4] === 'f' &&
-      addr[5] === 'f' &&
-      addr[6] === ':'
-    ) {
-      return true
-    } else if (isIP(addr.slice(2)) === 4) {
-      return true
-    }
-  }
-
-  let group = 0
-  for (var i = 0; i < addr.length; ++i) {
-    switch (addr[i]) {
-      case ':':
-        if (group === 5) {
-          return true
-        }
-        ++group
-        break
-      case '0':
-        if (group === 5) {
-          return false
-        }
-        break
-      case 'f':
-        if (group !== 5) {
-          return false
-        }
-        break
-      default:
-        return false
-    }
-  }
-
-  return false
-}
-
 /**
  * Module exports.
  * @public
@@ -452,5 +344,3 @@ module.exports.default = proxyaddr
 module.exports.proxyaddr = proxyaddr
 module.exports.all = alladdrs
 module.exports.compile = compile
-module.exports.prefixLengthFromSubnetMask = prefixLengthFromSubnetMask
-module.exports.isIPv4MappedIPv6Address = isIPv4MappedIPv6Address
